\name{income}
\alias{income}
\docType{data}
\title{income mean in 1000 $ on subsamples of age, gender and marital status groups}
\description{
  
}
\usage{data(income)}
\format{
  A data frame with 8 observations and 6 variables.
  \describe{
    \item{agegroup}{Age group}
    \item{maritalstatus}{Marital Status}
    \item{gender}{Gender}
    \item{samplesize}{Sample size in age group}
    \item{nresp}{Number of respondents in age group}
    \item{income}{Average income in group}
}
}

\source{Statistical Analysis with Missing data, Little and Rubin, 2nd Ed. p.~57}
\examples{
  data(cholesterol)
}
\keyword{datasets}
